move wallet modal
