+++ src/components/donation/Steps/DonateMethods/WalletModal/Connector.tsx
