+++ src/components/donation/Steps/DonateMethods/WalletModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./WalletModal";
